month2_minimart_analysis/python/requirements.txt
psycopg2
python-dotenv


